Avoid gcc warning on Red Hat 9:
secret.cpp:39: WARNING: cannot pass objects of non-POD type `class
   QByteRef' through `...'; call will abort at runtime

[git-p4: depot-paths = "//depot/qt/main/": change = 104195]

+++ examples/dragdrop/secret.cpp
@@ -36,7 +36,7 @@ bool SecretDrag::decode( QDropEvent* e, QString& str )
     if ( payload.size() ) {
 	e->accept();
 	QString msg;
-	msg.sprintf("The secret number is %d", payload[0] );
+	msg.sprintf("The secret number is %d", (int)payload[0] );
 	str = msg;
 	return TRUE;
     }
